Move controllers to coursewareable namespace.
